불필요한 추상화 제거 , 코드 복잡도 감소 위해 ProductService 는 인터페이스 없이 클래스로 구현하는것으로 변경
